+++ tests/testthat/test-likelihood.R
@@ -1,0 +1,19 @@
+# Tests of the likelihood-related functions
+# Arseniy Khvorov
+# Created 2019/10/17
+# Last edit 2019/10/17
+
+library(sclr)
+
+test_that("Likelihood works", {
+  fit <- sclr(status ~ logHI, one_titre_data)
+  expect_true(is.numeric(sclr_log_likelihood(fit)))
+  sclr_log_likelihood(fit, pars = c("lambda" = 0.5, "beta_0" = 1, "beta_1" = 1))
+  expect_true(is.numeric(sclr_log_likelihood(
+    fit, pars = c("lambda" = 0.5, "beta_0" = 1, "beta_1" = 1)
+  )))
+  expect_error(sclr_log_likelihood())
+  expect_error(
+    sclr_log_likelihood(list(x = fit$x, y = fit$y), pars = coef(fit))
+  )
+})